+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.3.2] - 2018-01-22
+### Changed
+- Action schema entity's primary key field name is getted from the
+  entity instead of the action.
+- Removed `ActionSchema.get_primary_key()`.
+- Binary values are logged as base64 encoded strings.
+
 ## [1.3.1] - 2017-11-03
 ### Added
 - A configured file server must be available to make runtime calls